requirements.txt
opencv-python
dlib
face_recognition
pillow
tk
pickle
pandas